fix: delete modal issue
